Move cljs-lib output - WIP
