+++ src/calva-fmt/src/infer.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { inferParens, inferIndents } from '../../../out/cljs-lib/cljs-lib';
+import { inferParens, inferIndents } from '../../cljs-lib/out/cljs-lib.js';
 import { isUndefined, cloneDeep } from 'lodash';
 
 interface CFEdit {
