+++ src/debugger/calva-debug.ts
@@ -34,7 +34,7 @@ import { moveTokenCursorToBreakpoint } from './util';
 import annotations from '../providers/annotations';
 import { NReplSession } from '../nrepl';
 import debugDecorations from './decorations';
-import { setStateValue, getStateValue } from '../../out/cljs-lib/cljs-lib';
+import { setStateValue, getStateValue } from '../cljs-lib/out/cljs-lib';
 import * as util from '../utilities';
 import * as replSession from '../nrepl/repl-session';
